% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModels.R
\name{SEI2HRModel}
\alias{SEI2HRModel}
\title{SEI2HRModel model}
\usage{
SEI2HRModel(initialConditionsQ = TRUE, rateRulesQ = TRUE, birthsTermQ = FALSE)
}
\arguments{
\item{initialConditionsQ}{Should initial conditions be included?}

\item{rateRulesQ}{Should rate rules be included?}

\item{birthsTermQ}{Should the births term be included or not?}

\item{A}{list}
}
\description{
SEI2HRModel
}
