% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModels.R
\name{SEI2RModel}
\alias{SEI2RModel}
\title{SEI2RModel model}
\usage{
SEI2RModel(initialConditionsQ = TRUE, rateRulesQ = TRUE, birthsTermQ = FALSE)
}
\arguments{
\item{initialConditionsQ}{Should initial conditions be given?}

\item{rateRulesQ}{Should rate rules be given?}

\item{birthsTermQ}{Should the births term be included or not?}

\item{A}{list}
}
\description{
SEI2RModel
}
