% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModels.R
\name{SI2RModel}
\alias{SI2RModel}
\title{SI2R model}
\usage{
SI2RModel(initialConditionsQ = TRUE, rateRulesQ = TRUE)
}
\arguments{
\item{initialConditionsQ}{Should initial conditions be given?}

\item{rateRulesQ}{Should rate rules be given?}

\item{A}{list}
}
\description{
SI2RModel
}
