% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpidemiologyModels.R
\name{SIRModel}
\alias{SIRModel}
\title{SIR model}
\usage{
SIRModel(initialConditionsQ = TRUE, rateRulesQ = TRUE)
}
\arguments{
\item{initialConditionsQ}{Should initial conditions be included?}

\item{rateRulesQ}{Should rate rules be included?}

\item{A}{list}
}
\description{
SIRModel
}
